requirements.txt
opencv-python
cvzone
Pillow
webbrowser
